refactor: move webpay transactions table to admin list table
